# src/main/resources/swagger/v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Mais informações sobre a API",
    "version": "1.0.0",
    "description": "API de Eventos feita com Quarkus \n - Autor: [Lucas S. Campos](https://www.linkedin.com/in/lucassantosc/) \n - Data: 06/04/2025 \n - Versão: 1.0.0 \n - Repositório: [API Eventos TSI](https://github.com/GiltMonster/API-Eventos-java) \n - ⚠️ **Atenção**: Esta versão da API ira se tornar depreciada em breve, por favor, utilize a versão mais recente da API. \n - ⚠️ **Atenção**: A V1 dessa API não possui autenticação, portanto, qualquer pessoa pode acessar os endpoints.",
    "termsOfService": "https://exemplo.com/termos",
    "contact": {
      "name": "Dev API - Support",
      "email": "lucas.scampos4@senacsp.edu.br"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "components": {
    "schemas": {
...
      "Date": {
        "type": "string",
        "format": "date",
        "examples": ["2022-03-10"]
      },
      "Evento": {
        "type": "object",
        "properties": {
          "evento_id": { "type": "integer", "format": "int64" },
          "nome": { "type": "string" },
          "descricao": { "type": "string" },
          "dataInicio": { "$ref": "#/components/schemas/Date" },
          "dataFim": { "$ref": "#/components/schemas/Date" },
          "localizacao": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/eventos/deletarEvento/{id}": {
      "delete": {
        "summary": "Deletar evento",
        "description": "Essa rota é responsável por deletar um evento do sistema.Se o evento for deletado com sucesso, retorna 200.Se o evento não for encontrado, retorna 404.",
        "parameters": [
          {
            "description": "ID do evento a ser deletado",
            "example": 1,
            "required": true,
            "name": "id",
            "in": "path",
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": { "description": "Evento deletado com sucesso" },
          "404": { "description": "Evento não encontrado" }
        },
        "tags": ["Evento Resource"]
      }
    },
    "/v1/eventos/findEvento/{id}": {
      "get": {
        "summary": "Buscar evento por ID",
        "description": "Essa rota é responsável por buscar um evento pelo ID.Se o ID não for encontrado, retorna 404.OBS: O ID é gerado automaticamente pelo banco de dados, é unico e é numérico.",
        "parameters": [
          {
            "description": "ID do evento a ser buscado",
            "example": 1,
            "required": true,
            "name": "id",
            "in": "path",
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado com sucesso",
            "content": {
              "application/json": {
                "examples": {
                  "Evento encontrado": {
                    "description": "Evento encontrado com sucesso",
                    "value": {
                      "evento_id": 1,
                      "nome": "Evento 1",
                      "descricao": "Descrição do evento 1",
                      "dataInicio": "2023-10-01T00:00:00Z",
                      "dataFim": "2023-10-02T00:00:00Z",
                      "localizacao": "Localização do evento 1"
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/Evento" }
              }
            }
          },
          "404": { "description": "Evento não encontrado" }
        },
        "tags": ["Evento Resource"]
      }
    },
...
  }
}